% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passt.R
\name{calc_output_sum}
\alias{calc_output_sum}
\title{Calculates sum of activation in output units for input patterns}
\usage{
calc_output_sum(inputs, weight_matrix)
}
\arguments{
\item{inputs}{matrix of inputs for which to calculate output
activation}

\item{weight_matrix}{weight matrix of network}
}
\value{
sum of activations of output units for input patterns
}
\description{
Calculates sum of activation in output units for input patterns
}
