% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passt.R
\name{create_pres_matrix}
\alias{create_pres_matrix}
\title{Create presentation matrix}
\usage{
create_pres_matrix(patterns, frequency, duration, lrate_onset,
  lrate_drop_time, lrate_drop_perc, pulses_per_second)
}
\arguments{
\item{patterns}{matrix with input patterns, one row is one pattern}

\item{frequency}{presentation frequency for each pattern in the
matrix}

\item{duration}{presentation duration for each pattern in the
matrix}

\item{lrate_onset}{learning rate at the onset of a stimulus}

\item{lrate_drop_time}{point at which the learning rate drops, must
be lower than duration}

\item{lrate_drop_perc}{how much the learning rate drops at
lrate_drop_time}

\item{pulses_per_second}{how many time steps should be simulated
per second}
}
\value{
list with two lists: the presentation matrix and the
  learning weights (both in one random order)
}
\description{
Create presentation matrix
}
