% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passt.R
\name{get_attention}
\alias{get_attention}
\title{Calculates attention (learning rate) development over time}
\usage{
get_attention(duration, lrate_onset, lrate_drop_time, lrate_drop_perc,
  pulses_per_second = 1)
}
\arguments{
\item{duration}{presentation duration for each pattern in the
matrix}

\item{lrate_onset}{learning rate at the onset of a stimulus}

\item{lrate_drop_time}{point at which the learning rate drops, must
be lower than duration}

\item{lrate_drop_perc}{how much the learning rate drops at
lrate_drop_time}

\item{pulses_per_second}{how many time steps should be simulated
per second}
}
\value{
list of attention values (learning rate) for every time
  pulse for every stimulus
}
\description{
Calculates attention (learning rate) development over time
}
