% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passt.R
\name{init_weight_mtrx}
\alias{init_weight_mtrx}
\title{Initialize weight matrix for competitive learning network. The
function draws from a normal distribution and then normalizes the
weights, such that the sum of weights is 1.}
\usage{
init_weight_mtrx(n_input_units, n_output_units, mean = 0.5, sd = 0.005)
}
\arguments{
\item{n_input_units}{number of input units}

\item{n_output_units}{number of output units}

\item{mean}{mean of normal distribution}

\item{sd}{sd of normal distribution}
}
\value{
mtrx with n_input_units rows and n_output_units columns,
  the sum of every column is 1
}
\description{
Initialize weight matrix for competitive learning network. The
function draws from a normal distribution and then normalizes the
weights, such that the sum of weights is 1.
}
