% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passt_doc.R
\docType{package}
\name{passt}
\alias{passt}
\alias{passt-package}
\title{passt: Simulates judgments of frequency and duration with a
competitive learning network based on the PASS-family}
\description{
To run a simulation you will just need to use the function
\code{\link[passt]{run_sim}}, which will take care of the rest. To
run several simulations and get a standard analysis use the
function \code{\link{run_exp}}. Both functions take care of the
details, while you are still able to set the most important
parameters such as the learning rate development, the input
patterns and the independent variables of exposure frequency and
exposure duration.
}
