% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passt.R
\name{updt_winner_weights}
\alias{updt_winner_weights}
\title{Updates weights for competitive learning network algorithm}
\usage{
updt_winner_weights(input, weight_matrix, lrate)
}
\arguments{
\item{input}{input vector}

\item{weight_matrix}{weight matrix of network}

\item{lrate}{learning rate}
}
\value{
new weight matrix for step t + 1
}
\description{
Updates weights for competitive learning network algorithm
}
