% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passt.R
\name{run_exp}
\alias{run_exp}
\title{Runs one "experiment" with several "participants", which means that
a simulation is run for each "participant" and correlative effect
sizes are calculated; stimuli are orthogonal}
\usage{
run_exp(duration, frequency, lrate_onset, lrate_drop_time, lrate_drop_perc,
  patterns = diag(length(duration)), number_of_participants = 100,
  cor_noise_sd = 0)
}
\arguments{
\item{duration}{presentation duration for each pattern in the
matrix}

\item{frequency}{presentation frequency for each pattern in the
matrix}

\item{lrate_onset}{learning rate at the onset of a stimulus}

\item{lrate_drop_time}{point at which the learning rate drops, must
be lower than duration}

\item{lrate_drop_perc}{how much the learning rate drops at
lrate_drop_time}

\item{patterns}{matrix with input patterns, one row is one pattern}

\item{number_of_participants}{corresponds with number of simulations run}

\item{cor_noise_sd}{the amount of noise added to the final activations of
the network, set to 0 if you do not want any noise}
}
\description{
Runs one "experiment" with several "participants", which means that
a simulation is run for each "participant" and correlative effect
sizes are calculated; stimuli are orthogonal
}
